+++ src/main/java/roomescape/auth/AuthException.java
@@ -7,7 +7,7 @@ public enum AuthException implements ExceptionCode {
 
     ACCESS_TOKEN_EXPIRED(HttpStatus.UNAUTHORIZED, "access token 만료", "access token 만료되었습니다."),
     INVALID_ACCESS_TOKEN(HttpStatus.UNAUTHORIZED, "access token 형식 오류", "access token이 유효하지 않습니다."),
-    ACCESS_TOKEN_NOT_FOUND(HttpStatus.BAD_REQUEST, "access token 없음", "access token이 쿠키에 존재하지 않습니다."),
+    UNAUTHENTICATED_REQUEST(HttpStatus.UNAUTHORIZED, "인증되지 않은 요청", "로그인이 필요합니다."),
     FORBIDDEN_ADMIN_ACCESS(HttpStatus.FORBIDDEN, "관리자 권한 없음", "관리자 권한이 필요합니다.")
     ;
