+++ src/main/java/roomescape/auth/LoginMember.java
@@ -1,6 +1,6 @@
 package roomescape.auth;
 
-import roomescape.member.Role;
+import roomescape.member.domain.Role;
 
 public record LoginMember(
         Long id,
