+++ src/main/java/roomescape/common/CookieUtil.java
@@ -5,7 +5,6 @@ import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.stereotype.Component;
 import roomescape.auth.AuthException;
 import roomescape.exception.ApplicationException;
-import roomescape.member.MemberException;
 
 @Component
 public class CookieUtil {
@@ -20,13 +19,13 @@ public class CookieUtil {
 
     public static String extractToken(Cookie[] cookies) {
         if (cookies == null) {
-            throw new ApplicationException(AuthException.ACCESS_TOKEN_NOT_FOUND);
+            throw new ApplicationException(AuthException.UNAUTHENTICATED_REQUEST);
         }
         for (Cookie cookie : cookies) {
             if ("token".equals(cookie.getName())) {
                 return cookie.getValue();
             }
         }
-        throw new ApplicationException(AuthException.ACCESS_TOKEN_NOT_FOUND);
+        throw new ApplicationException(AuthException.UNAUTHENTICATED_REQUEST);
     }
 }
