+++ src/main/java/roomescape/member/application/MemberService.java
@@ -3,10 +3,10 @@ package roomescape.member.application;
 import org.springframework.stereotype.Service;
 import roomescape.auth.JwtTokenProvider;
 import roomescape.exception.ApplicationException;
-import roomescape.member.Member;
-import roomescape.member.MemberDao;
-import roomescape.member.MemberException;
-import roomescape.member.Role;
+import roomescape.member.domain.Member;
+import roomescape.member.infrastructure.MemberDao;
+import roomescape.member.exception.MemberException;
+import roomescape.member.domain.Role;
 import roomescape.member.presentation.request.LoginRequest;
 import roomescape.member.presentation.request.MemberRequest;
 import roomescape.member.presentation.response.LoginMemberResponse;
