+++ src/main/java/roomescape/member/infrastructure/MemberDao.java
@@ -1,9 +1,11 @@
-package roomescape.member;
+package roomescape.member.infrastructure;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import roomescape.member.domain.Member;
+import roomescape.member.domain.Role;
 
 @Repository
 public class MemberDao {
