+++ src/main/java/roomescape/reservation/exception/ReservationException.java
@@ -1,4 +1,4 @@
-package roomescape.reservation;
+package roomescape.reservation.exception;
 
 import org.springframework.http.HttpStatus;
 import roomescape.exception.ExceptionCode;
