+++ src/main/java/roomescape/time/application/TimeService.java
@@ -1,11 +1,11 @@
 package roomescape.time.application;
 
 import org.springframework.stereotype.Service;
-import roomescape.reservation.Reservation;
-import roomescape.reservation.ReservationDao;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.infrastructure.ReservationDao;
 import roomescape.time.presentation.AvailableTime;
-import roomescape.time.Time;
-import roomescape.time.TimeDao;
+import roomescape.time.domain.Time;
+import roomescape.time.infrastructure.TimeDao;
 
 import java.util.List;
