+++ src/main/java/roomescape/time/presentation/TimeController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.time.Time;
+import roomescape.time.domain.Time;
 import roomescape.time.application.TimeService;
 
 import java.net.URI;
