+++ src/test/java/roomescape/MissionStepTest.java
@@ -7,7 +7,7 @@ import io.restassured.response.Response;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.reservation.presentation.ReservationResponse;
+import roomescape.reservation.presentation.response.ReservationResponse;
 
 import java.util.HashMap;
 import java.util.Map;
